% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvb.R
\name{qvb}
\alias{qvb}
\title{Generalized q - von Bertalanffy growth function
(Manabe et al. 2018. PLoS ONE 13(6):e0199346)}
\usage{
qvb(age, p = list(Lhat, r, q, tau, t_0))
}
\arguments{
\item{age}{Numerical age value(s) in year to estimate the length}

\item{p}{List of growth parameters
\describe{
\item{Lhat}{Growth parameter that set the size at inflection point}
\item{r}{Growth paramter that shapes the initial growth pattern}
\item{q}{Energy allocation parameter for reproduction.
Determinate growth is described when q < 1 and Indeterminate growth is
described when q > 1. In addition, q must not be 1.}
\item{tau}{Timing of maturity coefficient.}
\item{t_0}{Theoretical body size at 0}
}}
}
\value{
Estimated length at age in vector
}
\description{
Generalized q - von Bertalanffy growth function
(Manabe et al. 2018. PLoS ONE 13(6):e0199346)
}
\examples{
\dontrun{
qvb(age = seq(0, 5, 0.2), p = list(Lhat = 300, r = 0.75,
q = 1.2, tau = 2, t0 = -0.01))
}
}
