% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvb_gonad.R
\name{qvb_gonad}
\alias{qvb_gonad}
\title{Generalized q - von Bertalanffy growth function
(Manabe et al. 2018. PLoS ONE 13(6):e0199346)}
\usage{
qvb_gonad(L, p = list(Lhat, r, q, tau, c), a, b)
}
\arguments{
\item{L}{Length}

\item{p}{List of growth parameters
\describe{
\item{Lhat}{Growth parameter that set the size at inflection point}
\item{r}{Growth paramter that shapes the initial growth pattern}
\item{q}{Energy allocation parameter for reproduction.
Determinate growth is described when q < 1 and Indeterminate growth is
described when q > 1. In addition, q must not be 1.}
\item{tau}{Timing of maturity coefficient.}
\item{c}{Conversion factor that converts reproductive investment to GW}
}}

\item{a}{a of Length-weight relationship W = a * L ^ b}

\item{b}{b of Length-weight relationship W = a * L ^ b}
}
\value{
Estimated gonad weight
}
\description{
Generalized q - von Bertalanffy growth function
(Manabe et al. 2018. PLoS ONE 13(6):e0199346)
}
\examples{
\dontrun{
qvb_gonad(
L = 250,
p = list(300, 0.8, 1.2, 2, 0.005),
a = 0.0001,
b = 3.00)
}
}
